InstancingApp works (both platforms).  Transforms are messed up, need to fix.

+++ Engine/GraphicsDevice.h
@@ -34,7 +34,7 @@ public:
 	ColorBufferPtr GetBackBuffer(uint32_t index) const;
 	uint32_t GetCurrentBuffer() const { return m_currentBuffer; }
 
-	uint64_t GetFrameNumber() const { return m_frameNumber; }
+	uint32_t GetFrameNumber() const { return m_frameNumber; }
 
 	void ReleaseResource(PlatformHandle handle);
 
@@ -67,7 +67,7 @@ private:
 	std::array<ColorBufferPtr, NumSwapChainBuffers> m_swapChainBuffers;
 	uint32_t m_currentBuffer{ 0 };
 
-	uint64_t m_frameNumber{ 0 };
+	uint32_t m_frameNumber{ 0 };
 
 	// Deferred resource release
 	std::array<std::vector<PlatformHandle>, NumSwapChainBuffers> m_deferredReleasePages;
